HtmlBannerInAppDisplayProvider - custom-html header & footer notifications (#24)
Adds HtmlBannerInAppDisplayProvider to handle the display of custom-html header & footer banner-notifications.
The normal CleverTap implementation relies on a FragmentActivity which unreal can't easily support.

This functionality requires adding the CustomInAppDisplayProvider interface to the clevertap android sdk.
CleverTap/clevertap-android-sdk#839

This PR includes (and defaults to) a prebuilt clevertap-android-sdk.aar with this pre-release functionality included.
The AAR will be removed when the maven version has the needed features.